+++ internal/engine/session_internal_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/apex/log"
 	"github.com/google/go-cmp/cmp"
 	"github.com/ooni/probe-cli/v3/internal/checkincache"
+	"github.com/ooni/probe-cli/v3/internal/enginelocate"
 	"github.com/ooni/probe-cli/v3/internal/experiment/webconnectivity"
 	"github.com/ooni/probe-cli/v3/internal/experiment/webconnectivitylte"
-	"github.com/ooni/probe-cli/v3/internal/geolocate"
 	"github.com/ooni/probe-cli/v3/internal/kvstore"
 	"github.com/ooni/probe-cli/v3/internal/model"
 	"github.com/ooni/probe-cli/v3/internal/registry"
@@ -85,7 +85,7 @@ func TestSessionCheckInSuccessful(t *testing.T) {
 		},
 	}
 	s := &Session{
-		location: &geolocate.Results{
+		location: &enginelocate.Results{
 			ASN:         137,
 			CountryCode: "IT",
 		},
@@ -136,7 +136,7 @@ func TestSessionCheckInNetworkError(t *testing.T) {
 		Error: expect,
 	}
 	s := &Session{
-		location: &geolocate.Results{
+		location: &enginelocate.Results{
 			ASN:         137,
 			CountryCode: "IT",
 		},
@@ -178,7 +178,7 @@ func TestSessionCheckInCannotLookupLocation(t *testing.T) {
 func TestSessionCheckInCannotCreateProbeServicesClient(t *testing.T) {
 	errMocked := errors.New("mocked error")
 	s := &Session{
-		location: &geolocate.Results{
+		location: &enginelocate.Results{
 			ASN:         137,
 			CountryCode: "IT",
 		},
@@ -240,7 +240,7 @@ func TestSessionNewSubmitterWithCancelledContext(t *testing.T) {
 func TestSessionMaybeLookupLocationContextLookupLocationContextFailure(t *testing.T) {
 	errMocked := errors.New("mocked error")
 	sess := newSessionForTestingNoLookups(t)
-	sess.testLookupLocationContext = func(ctx context.Context) (*geolocate.Results, error) {
+	sess.testLookupLocationContext = func(ctx context.Context) (*enginelocate.Results, error) {
 		return nil, errMocked
 	}
 	err := sess.MaybeLookupLocationContext(context.Background())
