+++ internal/enginelocate/fake_test.go
@@ -1,4 +1,4 @@
-package geolocate
+package enginelocate
 
 import (
 	"net/http"
