+++ internal/enginelocate/geolocate_test.go
@@ -1,4 +1,4 @@
-package geolocate
+package enginelocate
 
 import (
 	"context"
